--- CONTRIBUTING.md
@@ -1,3 +1,0 @@
-# CONTRIBUTING
-
-Please don't open pull requests against this repository, please use https://github.com/zendframework/zf2.
