+++ EventManager.php
@@ -3,7 +3,7 @@
  * Zend Framework (http://framework.zend.com/)
  *
  * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2005-2014 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   http://framework.zend.com/license/new-bsd New BSD License
  */
 
@@ -530,7 +530,7 @@ class EventManager implements EventManagerInterface
      * Used to inject shared listeners and wildcard listeners.
      *
      * @param  PriorityQueue $masterListeners
-     * @param  array|Traversable $listeners
+     * @param  PriorityQueue $listeners
      * @return void
      */
     protected function insertListeners($masterListeners, $listeners)
