+++ SharedEventManager.php
@@ -3,7 +3,7 @@
  * Zend Framework (http://framework.zend.com/)
  *
  * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2005-2014 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   http://framework.zend.com/license/new-bsd New BSD License
  */
 
@@ -109,7 +109,7 @@ class SharedEventManager implements
     /**
      * Detach a listener aggregate
      *
-     * Listener aggregates accept a SharedEventManagerInterface instance, and call detachShared()
+     * Listener aggregates accept an SharedEventManagerInterface instance, and call detachShared()
      * of all previously attached listeners.
      *
      * @param  SharedListenerAggregateInterface $aggregate
