+++ composer.json
@@ -6,7 +6,6 @@
         "zf2",
         "eventmanager"
     ],
-    "homepage": "https://github.com/zendframework/zf2",
     "autoload": {
         "psr-0": {
             "Zend\\EventManager\\": ""
@@ -14,13 +13,13 @@
     },
     "target-dir": "Zend/EventManager",
     "require": {
-        "php": ">=5.3.23",
+        "php": ">=5.3.3",
         "zendframework/zend-stdlib": "self.version"
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "2.3-dev",
-            "dev-develop": "2.4-dev"
+            "dev-master": "2.2-dev",
+            "dev-develop": "2.3-dev"
         }
     }
 }
